requirement.txt
beautifulsoup4==4.9.3
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
cycler==0.10.0
Cython==0.29.14
decorator==4.4.2
et-xmlfile==1.0.1
fuzzywuzzy==0.18.0
gensim==3.8.3
graphviz==0.16
idna==2.10
isodate==0.6.0
jdcal==1.4.1
joblib==1.0.0
kiwisolver==1.3.1
lxml==4.6.2
matplotlib==3.3.4
networkx==2.5
ngram==3.3.2
nltk==3.5
numpy==1.19.5
openpyxl==3.0.6
Owlready2==0.26
pandas==1.1.5
Pillow==8.1.0
py-stringmatching==0.4.1
pydotplus==2.0.2
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
rdflib==5.0.0
regex==2020.11.13
requests==2.25.1
scikit-learn==0.24.1
scipy==1.5.4
six==1.15.0
sklearn==0.0
smart-open==4.1.2
soupsieve==2.1
stop-words==2018.7.23
threadpoolctl==2.1.0
tqdm==4.56.0
urllib3==1.26.3
wincertstore==0.2
xgboost==1.3.3
xlrd==2.0.1
XlsxWriter==1.3.7
